gen3_use_requirements.txt
numpy==1.24.3 
ultralytics==8.2.90
opencv-python==4.7.0.72
matplotlib==3.5.3
scikit-image==0.19.3
imageio==2.28.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
# tensorflow-estimator==2.5.0
# tensorflow-gpu==2.5.0
# torch==1.10.1+cu111
# torchaudio==0.10.1+cu111
torchsummary==1.5.1
# torchvision==0.11.2+cu111
pyyaml
imagecodecs
   